02/28/2025 - Finished
- Added Prefix
- Reimplemented SimplifyEquation method

The assignment wanted to use a stack, but without calculating a number solution it feels pointless, so I didn't. Also, It technically is not doing 'proper' postfix form, but it if you were to read the solution it does simplify to the same equation. I didn't really understand the actual format...